fix: 필드 타입 수정

+++ src/main/java/kr/hhplus/be/server/domain/order/command/OrderCommand.java
@@ -19,7 +19,7 @@ public record OrderCommand(
         long productId,
 
         @Schema(description = "상품 수량")
-        long quantity
+        int quantity
     ) {
     }
 }
