+++ src/main/java/kr/hhplus/be/server/interfaces/api/order/OrderRequest.java
@@ -29,7 +29,7 @@ public record OrderRequest(
         long productId,
 
         @Schema(description = "상품 수량")
-        long quantity
+        int quantity
 
     ) {
         public OrderItemCommand toCommand() {
